Remove tmux
tmux configuration was starting to bother me, and it got worse with
Sierra. Turns out that neovim terminal emulation is much more natural to
work with and requires less configuration, though I still disapprove of
the feature being included in an editor!

For detaching and attaching sessions, abduco is a simple tool that seems
to work decently, but given that I will likely only be running neovim in
the sessions, I might come up with something even simpler.