Clean up fbclock
You can't use the return value of gzerror to check if an error occurred
or not. Its implementation actually checks if the internal error is NULL
and returns the empty string if it is! This is stupid and unhelpful, so
check gzeof first since its return value actually means something.