Rearrange kshrc, fix right prompt
Including the \r in the delimeters works properly in every version of
ksh I have. Awesome.

Also switch to emacs editing mode because it works better and I know how
to use it now. And set CDPATH with an empty item so it doesn't print the
expansion all the time.