Add The Invisible Life of Addie LaRue
So good, but so long. Reminded me of The Ten Thousand Doors of
January at the beginning, and more of that N. K. Jemisin series
about gods later. I like this interacting with gods and becoming
something like one sort of thing.

God, it took me a whole month (more?) to read and this is only my
third book of the year :( I need some more novellas to read, but
the other books I have from the library currently are also thick.