Add Conversations With Friends
The better of the two, but largely the same. I feel like these books
are a bit too autobiographical, but I don't know if I'm allowed to
accuse an author of that.

My real problem is that I read these books as largely uncritical
of their characters. They behave in nonsense ways, are mostly
uncritical of their own behaviour, and don't really have arcs of
growth or change. I suppose this book had a bit of one, but only
in the last two chapters.