Add Tea with the Black Dragon
I found this in a box on the side of the road being thrown own.
It's decent pulp but nothing special. Not even as racist as I
expected it to be.

Also as one might guess, the author using their initials is a woman,
so I only broke one of my rules when it comes to book selection.